requirements.txt
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
bcrypt==4.2.0
psycopg2-binary==2.9.10
email-validator==1.1.3
Flask-WTF==1.0.1
#greenlet==1.1.2
#idna==3.3
#importlib-metadata==4.11.3
#itsdangerous==2.1.2
#Jinja2==3.1.1
#MarkupSafe==2.1.1
#pycparser==2.21
#six==1.16.0
#SQLAlchemy==1.4.35
#sqlparse==0.4.2
#Werkzeug==2.1.1
WTForms==3.0.1
python-dotenv==0.21.0
gunicorn==22.0.0

